Refactoring code to extract common methods to a module Adding tests to module too
